# [zmath] Replace swizzles with shuffles & remove some unnecessary math complexity to increase perf.