These changes fixes Issue #636 by changing swizzles to the builtin `@shuffle`, which generates smaller code.

There is a chance that the zig compiler is able to eventually fully optimize a swizzle call - but that isn't the case right now.

Other changes:
- `dot2` and `dot4` have been simplified
- `all` and `any` now use `@reduce` appropriately (which should offer SIMD speed improvements) as a comptime decision, and actually support float types now by falling back to looping.
   -  (added tests for float support) 

Perf results from M1 mac:
```
                matrix mul benchmark (AOS) - scalar version: 1.0043s, zmath version: 0.9783s
       cross3, scale, bias benchmark (AOS) - scalar version: 0.6268s, zmath version: 0.6478s
 cross3, dot3, scale, bias benchmark (AOS) - scalar version: 0.9808s, zmath version: 0.9543s
            quaternion mul benchmark (AOS) - scalar version: 0.9863s, zmath version: 0.7783s
                      wave benchmark (SOA) - scalar version: 3.4083s, zmath version: 1.0393s
```

(notice how the ` cross3, dot3, scale, bias benchmark` benchmark is now faster with zmath). Other benchmarks seem faster too, but it's hard to fully know.