@michal-z 2024-07-15
`swizzle()` is a trivial wrapper for `shuffle()` it shouldn't change anything. Do you observe a difference in code gen?