@hazeycode 2024-07-15
Changes look good to me but I'm not seeing any statistically significant improvement in the benchmarks (M3 Pro macOS 14.5).

I think some of the benchmarks are cache bound.