@hazeycode 2024-07-15
M3 Pro, MacOS 14.5 results:

#### `-Doptimize=Debug`
|      Benchmark     |    Before (s)    |    After (s)   |
|------------------| ---------------| --------------|
|                        matrix mul benchmark (AOS) | 32.6926  | 26.2363 | 
|           cross3, scale, bias benchmark (AOS) | 12.9573   | 10.0015  |
| cross3, dot3, scale, bias benchmark (AOS) | 14.2043   | 11.5713  |
|                quaternion mul benchmark (AOS)  | 20.4624  | 14.0007 |
|                                 wave benchmark (SOA)  |  6.5604   | 6.9920  |


#### `-Doptimize=ReleaseFast`
|      Benchmark     |    Before (s)    |    After (s)   |
|------------------| ---------------| --------------|
|                        matrix mul benchmark (AOS) | 0.7679  | 0.7680 | 
|           cross3, scale, bias benchmark (AOS) | 0.4885   | 0.5564  |
| cross3, dot3, scale, bias benchmark (AOS) | 0.8333  | 0.8264  |
|                quaternion mul benchmark (AOS)  | 0.6551  | 0.6535 |
|                                 wave benchmark (SOA)  | 0.7284  | 0.7302  |