@dmurph 2024-07-15
I tried to create a minimal godbolt to show the generated output
https://godbolt.org/z/8oP6v7jvx
(you'll have to search for the methods in the generated output)

What's interesting here now is that it is now able to fully optimize the swizzle in the `main` function. In the old example using the [dot4Old function in this godbolt](https://godbolt.org/z/EfM5b5qs9) - by swapping swizzle for `@shuffle` for it generated less instructions.

Anyways - the dot4, dot2, any, and all are clearly much better instruction-wise. I can remove the `@shuffle` changes if you want.