@hazeycode 2024-07-15
`@shuffle` looks good to me!

Overall the improvements are very obvious in Debug mode benchmarks.