@michal-z 2024-07-17
@dmurph You must use `-O ReleaseFast` in godbolt to enable optimized build (-DReleaseFast doesn't work and produces debug code).