@dmurph 2024-07-17
Very cool! thanks.


Ok, after that, here are the new things for dots:

dot4Old:
```
        vmulps  xmm0, xmm0, xmmword ptr [rsp + 16]
        vshufps xmm1, xmm0, xmm0, 177
        vaddps  xmm0, xmm0, xmm1
        vshufps xmm1, xmm0, xmm0, 27
        vaddps  xmm0, xmm0, xmm1
```

dot4:
```
        vmulps  xmm0, xmm0, xmmword ptr [rsp + 16]
        vmovshdup       xmm1, xmm0
        vaddss  xmm1, xmm0, xmm1
        vshufpd xmm2, xmm0, xmm0, 1
        vaddss  xmm1, xmm2, xmm1
        vshufps xmm0, xmm0, xmm0, 255
        vaddss  xmm0, xmm0, xmm1
        vbroadcastss    xmm0, xmm0
```

dot2Old:
```
        vmulps  xmm0, xmm0, xmmword ptr [rsp + 16]
        vmovshdup       xmm1, xmm0
        vaddss  xmm0, xmm0, xmm1
        vbroadcastss    xmm0, xmm0
```

dot2:
```
        vmulps  xmm0, xmm0, xmmword ptr [rsp + 16]
        vmovshdup       xmm1, xmm0
        vaddss  xmm0, xmm0, xmm1
        vbroadcastss    xmm0, xmm0
```

The shuffle vs swizzle also fully optimizes to be the same thing.

I wrote incorrect code for  .all and .any - I'm going to fix that up and see what the difference is.

Here is my current compiler explorer: https://godbolt.org/z/7E9YW8oqv