@dmurph 2024-07-18
Cool - after fixing the `all` function, I can say that that certainly improved instructions:

before, len 3 of size 4 vector - 10
```
        cmp     dword ptr [rsp + 64], 0
        setne   cl
        cmp     dword ptr [rsp + 68], 0
        setne   dl
        and     dl, cl
        cmp     dword ptr [rsp + 72], 0
        setne   cl
        and     cl, dl
        mov     byte ptr [rsp + 11], cl
        lea     rcx, [rsp + 11]
```

after, worst case - 7
```
        vpbroadcastq    xmm0, qword ptr [rsp + 72]
        vpand   xmm0, xmm0, xmmword ptr [rsp + 64]
        vpsrlq  xmm1, xmm0, 32
        vpand   xmm0, xmm0, xmm1
        vmovd   eax, xmm0
        test    eax, eax
        setne   byte ptr [rsp + 11]
```

(When len=3) - 4
```
        mov     eax, dword ptr [rsp + 64]
        and     eax, dword ptr [rsp + 68]
        test    dword ptr [rsp + 72], eax
        setne   byte ptr [rsp + 11]
```

setting the cpu to x86_64 - 8
```
        movdqa  xmm0, xmmword ptr [rsp + 64]
        pshufd  xmm1, xmm0, 238
        movd    eax, xmm1
        pshufd  xmm0, xmm0, 85
        movd    edx, xmm0
        and     edx, dword ptr [rsp + 64]
        test    edx, eax
        setne   byte ptr [rsp + 11]
```

https://godbolt.org/z/d9W4hMco9

I'm going to remove the dot changes and keep the 'any' and 'all' changes. Let me know if you want me to keep the shuffle changes, as they seem to affect debug builds but not release.