@dmurph 2024-07-18
I guess let me know if you want any of this - happy to just close this request as my original changes weren't actually more performant lol.