@hazeycode 2024-07-20
Debug perf is important and we should consider changes to improve it carefully.