@dmurph 2024-07-22
I've gotten a bit busy - some thoughts:

- I'll plan on splitting this up to discuss separately
    - all/any change (not sure if that's wanted) adds 'fast' support for int & bool calls to that. Float is identical. This 'new' support is probably not really important / needed. Happy to abandon that.
- swizzle -> shuffle change in zmath.zig to help debug builds

Since this won't be for a bit - feel free to take over this patch if you like if you feel inspired or have free time. Otherwise I'll likely revisit in a week or so.